+++ docs/about/release-notes.md
@@ -16,6 +16,7 @@
 - GetCube: add CompressionLevel=-3 to disable the compression
 - GetCube: add Predownload option to download file before warping to save time. It is efficient when gdal needs the whole image to compute the Cube requested, but its not when a small part of the image is required. Be careful when the data has been consolidated.
 - Consolidation: add collapse_on_record_id: to consolidate by collapsing all datasets on the given record (data is copied)
+- Get information on containers from their uris
 
 ### Bug fixes
 - CleanJobs: remove DONEBUTUNTIDY